# api/openapi.yaml
openapi: 3.1.0
info:
  title: nav.ax API
  version: 1.0.0
  summary: nav.ax 导航站 v1 前后端唯一接口契约
  description: |
    所有时间均为 UTC RFC 3339，所有 ID 均为不透明字符串。
    认证采用 HttpOnly Session Cookie；认证后的非安全方法还必须通过 Origin 校验。
    成功响应统一为 `{code,data,meta}`，错误响应的 `data` 固定为 null。
servers:
  - url: /
...
paths:
...
  /api/v1/me/subdomain:
    get:
      tags: [Pages]
      operationId: getMySubdomain
      security: [{ sessionCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/SubdomainResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
    post:
      tags: [Pages]
      operationId: requestSubdomain
      description: 申请 nav.ax 子域名。4 个及以上字符在名称可用时自动通过；1 至 3 个字符作为稀缺短域名进入管理员审核。
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubdomainRequestInput' }
      responses:
        '201': { $ref: '#/components/responses/SubdomainResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
    patch:
      tags: [Pages]
      operationId: setSubdomainCustomDomain
      description: 为已批准子域名绑定可选 CNAME 主机名（需在 DNS 侧将 CNAME 指向 fullDomain）
      security: [{ sessionCookie: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                customDomain: { type: [string, 'null'], maxLength: 253, description: 空或 null 表示清除 }
      responses:
        '200': { $ref: '#/components/responses/SubdomainResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
    delete:
      tags: [Pages]
      operationId: cancelSubdomainRequest
      security: [{ sessionCookie: [] }]
      responses:
        '200': { $ref: '#/components/responses/EmptyResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/admin/subdomains:
    get:
      tags: [Admin]
      operationId: listSubdomainRequests
      security: [{ sessionCookie: [] }]
      parameters:
        - in: query
          name: status
          schema: { $ref: '#/components/schemas/SubdomainStatus' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200': { $ref: '#/components/responses/SubdomainsResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/admin/subdomains/{requestId}:
    patch:
      tags: [Admin]
      operationId: reviewSubdomainRequest
      security: [{ sessionCookie: [] }]
      parameters:
        - in: path
          name: requestId
          required: true
          schema: { $ref: '#/components/schemas/Id' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SubdomainReviewRequest' }
      responses:
        '200': { $ref: '#/components/responses/SubdomainResponse' }
        '401': { $ref: '#/components/responses/ErrorResponse' }
        '403': { $ref: '#/components/responses/ErrorResponse' }
        '409': { $ref: '#/components/responses/ErrorResponse' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: navax_session
  parameters:
...
    Page: { in: query, name: page, schema: { type: integer, minimum: 1, default: 1 } }
    PageSize: { in: query, name: pageSize, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
...
  responses:
    ErrorResponse:
      description: 统一错误响应
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    EmptyResponse: { description: 操作成功, content: { application/json: { schema: { $ref: '#/components/schemas/EmptyEnvelope' } } } }
...
    SubdomainResponse: { description: 子域名申请；无申请时 data 为 null, content: { application/json: { schema: { $ref: '#/components/schemas/SubdomainEnvelope' } } } }
    SubdomainsResponse: { description: 子域名申请列表, content: { application/json: { schema: { $ref: '#/components/schemas/SubdomainsEnvelope' } } } }
...
  schemas:
    Id: { type: string, minLength: 8, maxLength: 64 }
...
    Timestamp: { type: string, format: date-time }
    ResponseMeta:
      type: object
      required: [requestId]
      properties:
        requestId: { type: string }
        message: { type: string }
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
    ErrorMeta:
      allOf:
        - $ref: '#/components/schemas/ResponseMeta'
        - type: object
          required: [message]
          properties:
            message: { type: string }
            detail: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
    ErrorEnvelope:
      type: object
      required: [code, data, meta]
      properties:
        code: { type: string, enum: [MALFORMED_REQUEST, VALIDATION_FAILED, AUTH_REQUIRED, INVALID_CREDENTIALS, SESSION_EXPIRED, FORBIDDEN, ADMIN_REQUIRED, ACCOUNT_DISABLED, NOT_FOUND, METHOD_NOT_ALLOWED, INSTANCE_ALREADY_INITIALIZED, EMAIL_CONFLICT, USERNAME_CONFLICT, SLUG_CONFLICT, CATEGORY_NOT_EMPTY, INVITATION_EXPIRED, INVITATION_REVOKED, INVITATION_EXHAUSTED, INVITATION_NOT_FOUND, INVALID_RESET_TOKEN, DRAFT_REVISION_MISMATCH, IDEMPOTENCY_CONFLICT, IMPORT_TOO_LARGE, IMPORT_CONFLICT, IMPORT_PREVIEW_EXPIRED, INVALID_IMPORT, INVALID_URL, IMPORT_INVALID, PAYLOAD_TOO_LARGE, UNSUPPORTED_MEDIA_TYPE, LINK_CHECK_BUSY, RATE_LIMITED, SETUP_REQUIRED, SERVICE_UNAVAILABLE, CONFLICT, INTERNAL_ERROR] }
        data: { type: 'null' }
        meta: { $ref: '#/components/schemas/ErrorMeta' }
    EmptyEnvelope:
      type: object
      required: [code, data, meta]
      properties:
        code: { const: OK }
        data: { type: 'null' }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
...
    SubdomainStatus: { type: string, enum: [pending, approved, rejected, revoked] }
    Subdomain:
      type: object
      description: 子域名记录。标准子域名可由系统自动批准，此时 reviewedAt 为自动启用时间。
      required: [id, userId, label, fullDomain, status, appliedAt]
      properties:
        id: { $ref: '#/components/schemas/Id' }
        userId: { $ref: '#/components/schemas/Id' }
        username: { type: string }
        label: { type: string }
        fullDomain: { type: string }
        customDomain: { type: [string, 'null'], description: 可选 CNAME 主机名 }
        status: { $ref: '#/components/schemas/SubdomainStatus' }
        appliedAt: { $ref: '#/components/schemas/Timestamp' }
        reviewedAt: { oneOf: [{ $ref: '#/components/schemas/Timestamp' }, { type: 'null' }] }
        reason: { type: string }
    SubdomainRequestInput:
      type: object
      required: [label]
      properties:
        label:
          type: string
          description: 1 至 30 个小写字母、数字或连字符；不能以连字符开头或结尾。4 个及以上字符自动通过，1 至 3 个字符需审核。
          pattern: '^[a-z0-9](?:[a-z0-9-]{0,28}[a-z0-9])?$'
          minLength: 1
          maxLength: 30
    SubdomainReviewRequest:
      type: object
      required: [decision]
      properties:
        decision: { type: string, enum: [approve, reject, revoke] }
        reason: { type: string, maxLength: 300 }
...
    SubdomainEnvelope: { allOf: [{ $ref: '#/components/schemas/EnvelopeBase' }, { type: object, properties: { data: { oneOf: [{ $ref: '#/components/schemas/Subdomain' }, { type: 'null' }] } }, required: [data] }] }
    SubdomainsEnvelope: { allOf: [{ $ref: '#/components/schemas/PaginatedEnvelopeBase' }, { type: object, properties: { data: { type: array, items: { $ref: '#/components/schemas/Subdomain' } } }, required: [data] }] }
...
    EnvelopeBase:
      type: object
      required: [code, meta]
      properties:
        code: { const: OK }
        meta: { $ref: '#/components/schemas/ResponseMeta' }
    PaginatedEnvelopeBase:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            meta:
              allOf:
                - $ref: '#/components/schemas/ResponseMeta'
                - type: object
                  required: [page, pageSize, total, totalPages]
                  properties:
                    page: { type: integer, minimum: 1 }
                    pageSize: { type: integer, minimum: 1 }
                    total: { type: integer, minimum: 0 }
                    totalPages: { type: integer, minimum: 0 }